pyproject.toml
[tool.poetry]
name = "kmibot"
version = "0.1.0"
description = ""
authors = ["Dan Trickey <git@trickey.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^2.1.0"
pydantic = "^1.10.2"
tomli = "^2.0.1"
dateparser = "^1.1.8"

[tool.poetry.group.dev.dependencies]
ruff = "*"
mypy = "*"
types-dateparser = "^1.1.4.9"

[tool.ruff]
target-version = "py39"
select = [
    "A",
    "ANN",
    "B",
    "BLE",
    "C4",
    "COM",
    "DJ",
    "DTZ",
    "E",
    "F",
    "FBT",
    "I",
    "N",
    "S",
    "T10",
    "UP",
    "W",
]
ignore = [
    "ANN101",  # Missing type annotation for `self` in method
    "ANN401",  # Dynamically typed expressions (typing.Any) are disallowed
    "B009",    # Do not call `getattr` with a constant attribute value.
    "S101",    # S101 Use of `assert` detected
]

line-length = 120

[tool.mypy]
python_version = "3.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
